TM-8938 Changes after PR comments

+++ grails-app/services/net/transitionmanager/service/ImportBatchService.groovy
@@ -1,8 +1,8 @@
 package net.transitionmanager.service
 
 import com.tdsops.tm.enums.domain.ImportBatchStatusEnum
-import com.tdssrc.grails.GormUtil
 import groovy.util.logging.Slf4j
+import net.transitionmanager.command.IdsCommandObject
 import net.transitionmanager.domain.ImportBatch
 import net.transitionmanager.domain.Project
 import net.transitionmanager.i18n.Message
@@ -28,34 +28,6 @@ class ImportBatchService implements ServiceMethods{
 		}.list()
 	}
 
-	/**
-	 * Find a batch for the given project with the id provided.
-	 *
-	 * @param importBatchId - the id of the batch.
-	 * @param project - the project
-	 * @return the batch, null if it doesn't exist.
-	 */
-	ImportBatch findImportBatchById(Long importBatchId, Project project) {
-		return GormUtil.findInProject(project, ImportBatch, importBatchId, true)
-	}
-
-	/**
-	 * Delete a batch if it exists for this project, throwing an exception if it doesn't
-	 * or if it's being processed.
-	 *
-	 * @param importBatchId - the batch id
-	 * @param project
-	 */
-	void deleteImportBatch(Long importBatchId, Project project) {
-		ImportBatch importBatch = findImportBatchById(importBatchId, project)
-		if (! importBatch) {
-			throw new EmptyResultException(Message.ImportBatchDoesntExist)
-		}
-		if (importBatch.status == ImportBatchStatusEnum.RUNNING) {
-			throw new EmptyResultException(Message.ImportBatchRunning)
-		}
-		importBatch.delete()
-	}
 
 	/**
 	 * Delete a list of Import Batches, throwing an exception if not all of them
@@ -63,55 +35,48 @@ class ImportBatchService implements ServiceMethods{
 	 * @param importBatchIds
 	 * @param project
 	 */
-	void bulkDeleteImportBatches(List<Long> importBatchIds, Project project) {
-		/* Even though the ids are received as a List<Long>, internally they are handled as Integers in the query,
-		causing a casting exception. */
-		String idsList = importBatchIds.join(", ")
-		String hql = """
+	void deleteImportBatch(IdsCommandObject idsCommand, Project project) {
+		if (idsCommand.validate()) {
+			List<Long> importBatchIds = idsCommand.ids
+			// Query that will delete all not running jobs for this project from a list of ids.
+			String hql = """
 				DELETE FROM ImportBatch 
-				WHERE project = :project AND id in ($idsList) AND status != :status"""
-		Map hqlParams = [project: project, status: ImportBatchStatusEnum.RUNNING]
-		int deletedBatches = ImportBatch.executeUpdate(hql, hqlParams)
-		// Throw an exception if not all the batches were deleted
-		if (importBatchIds.size() > deletedBatches) {
-			throw new EmptyResultException(Message.ImportBatchBulkDelete)
+				WHERE project = :project AND id in (:batches) AND status != :status"""
+			Map hqlParams = [project: project, status: ImportBatchStatusEnum.RUNNING, batches: importBatchIds]
+			// Execute the query up keep and retrieve the number of records deleted.
+			int deletedBatches = ImportBatch.executeUpdate(hql, hqlParams)
+			// Throw an exception if not all the batches were deleted
+			if (importBatchIds.size() > deletedBatches) {
+				throw new DomainUpdateException(Message.ImportBatchBulkDelete)
+			}
+		} else {
+			throw new InvalidRequestException(Message.RequestMissingIds)
 		}
+
 	}
 
 	/**
-	 * Set the archived property's value on the Import Batch to that of the parameter received.
-	 * This method will throw an EmptyResultException if the given import batch doesn't exist.
+	 * Set the archived property's value on the Import Batches received.
 	 * @param importBatchId - the id of the batch
 	 * @param project - current project
 	 * @param archived - true: archive / false: unarchive
 	 */
-	void setArchivedFlagOnImportBatch(Long importBatchId, Project project, boolean archived) {
-		ImportBatch importBatch = findImportBatchById(importBatchId, project)
-		if (! importBatch) {
-			throw new EmptyResultException(Message.ImportBatchDoesntExist)
+	void setArchivedFlagOnImportBatch(IdsCommandObject idsCommand, Project project, boolean archivedFlag) {
+		if (idsCommand.validate()) {
+			List<Long> importBatchIds = idsCommand.ids
+			// Transform the boolean flag to 0: false, 1: true
+			Integer archived = BooleanUtils.toInteger(archivedFlag)
+			// Query that archives or unarchives a list of batches.
+			String hql = "UPDATE ImportBatch SET archived = :archived WHERE project = :project AND id in (:batches)"
+			Map params = [archived: archived, project: project, batches: importBatchIds]
+			Integer updated = ImportBatch.executeUpdate(hql, params)
+			// Throw an exception if the number of updated records doesn't match the number of batches received.
+			if (updated < importBatchIds.size()) {
+				throw new DomainUpdateException(Message.ImportBatchBulkUpdate)
+			}
+		} else {
+			throw new InvalidRequestException(Message.RequestMissingIds)
 		}
-		importBatch.archived = BooleanUtils.toInteger(archived)
-		importBatch.save(failOnError: true)
-	}
 
-	/**
-	 * Set the 'archived' flag on a list of import batches.
-	 * @param importBatchIds
-	 * @param project
-	 * @param archivedFlag
-	 */
-	void bulkSetArchivedFlagOnImportBatches(List<Long> importBatchIds, Project project, boolean archivedFlag) {
-		// Transform the boolean flag to 0: false, 1: true
-		Integer archived = BooleanUtils.toInteger(archivedFlag)
-		/* Even though the ids are received as a List<Long>, internally they are handled as Integers in the query,
-		causing a casting exception. */
-		String idsList = importBatchIds.join(", ")
-		String hql = "UPDATE ImportBatch SET archived = :archived WHERE project = :project AND id in ($idsList)"
-		Map params = [archived: archived, project: project]
-		Integer updated = ImportBatch.executeUpdate(hql, params)
-		// Throw an exception if the number of updated records doesn't match the number of batches received.
-		if (updated < importBatchIds.size()) {
-			throw new EmptyResultException(Message.ImportBatchBulkUpdate)
-		}
 	}
 }
